Creating import_functional_only_matrices.R
The import code works fine for the matrices; an .RData file is created and looks correct. I "out-commented" a few lines in the code above that are part of the code for the metadata. The "original" HOCOMOCO file from their website that I used for the upload is called "HOCOMOCOv9_AD_PLAINTEXT_H_WPCM.txt".